Move frame_to_tensor to preprocessing
